Issues 356 - Moving enclosure logic from main to init
